Add files via upload
he current research presents an Artificial Neural Network (ANN) architecture which employs a Fully Connected Non-linear Network (FCNN) as the feature selector and an adaptive Radial Basis Function Network (RBFN) as the composure approximator of FCNN, for the application of fault detection in ventilation fan systems. In this approach, raw data is first preprocessed and then multiplied by an Adjustable Weight matrix and is activated using Sigmoid Linear Unit (Si-LU), after which is multiplied by yet another trainable Weight matrix and multiplied element-wise with the original preprocessed input to create the output of the FCNN, which has the same dimensionality as the input. The resulting output of the FCNN is then passed through the RBFN, which. Subsequently, a learnable exponential RBF layer, with trainable centers and standard deviations, transforms these features. The resulting features are mapped into a 2-dimensional latent space, and a linear classifier performs the final fault classification. Experimental outcomes show that this method successfully reduces the feature space from 10 dimensions to 2, surpassing traditional dimensionality reduction techniques like SVD, PCA, and nonlinear methods such as t-SNE. The approach achieves a macro F1-score of 97% on the training set and 95% on the test set, while also reducing data volume and computational runtime. Experiments also support precise fault classification and offers an interpretable latent space analysis.